redesigned the project section. Now there is no tab, rather 6 projects all wrap displayed on the page. all are an image with an overlay which displays the project info
